# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IRMS 后端 API（Gin + Gorm + Swagger）",
        "title": "IRMS API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/pages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pages"
                ],
                "summary": "列出页面资源",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (\u003e=1); invalid -\u003e 400",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (1~100); invalid -\u003e 400",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword (name like)",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "type": "string",
                        "description": "Status; invalid -\u003e 400",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.PageListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pages"
                ],
                "summary": "创建页面资源",
                "parameters": [
                    {
                        "description": "PageCreateRequest",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.PageCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.IDResponse"
                        }
                    }
                }
            }
        },
        "/pages/sync": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pages"
                ],
                "summary": "同步页面路由",
                "parameters": [
                    {
                        "description": "PageSyncRequest",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.PageSyncRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.PageSyncResponse"
                        }
                    }
                }
            }
        },
        "/pages/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pages"
                ],
                "summary": "更新页面资源",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PageUpdateRequest",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.PageUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.EmptyResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pages"
                ],
                "summary": "删除页面资源",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.EmptyResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "details": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "dtoresp.EmptyData": {
            "type": "object"
        },
        "dtoresp.EmptyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.EmptyData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "dtoresp.IDData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "dtoresp.IDResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.IDData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "dtoresp.PageListData": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.PageVO"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dtoresp.Pagination"
                }
            }
        },
        "dtoresp.PageListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.PageListData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "dtoresp.PageSyncResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/vo.PageSyncVO"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "dtoresp.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "request.PageCreateRequest": {
            "type": "object",
            "required": [
                "name",
                "route_path"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "route_path": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "request.PageSyncRequest": {
            "type": "object",
            "properties": {
                "dry_run": {
                    "type": "boolean"
                },
                "routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.PageSyncRouteItem"
                    }
                }
            }
        },
        "request.PageSyncRouteItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "route_path": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "request.PageUpdateRequest": {
            "type": "object",
            "required": [
                "name",
                "route_path",
                "status"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "route_path": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "vo.PageSyncRouteVO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "route_path": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "vo.PageSyncSummaryVO": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "dry_run": {
                    "type": "boolean"
                },
                "input_total": {
                    "type": "integer"
                },
                "retired": {
                    "type": "integer"
                },
                "unchanged": {
                    "type": "integer"
                },
                "updated": {
                    "type": "integer"
                }
            }
        },
        "vo.PageSyncVO": {
            "type": "object",
            "properties": {
                "changed_routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.PageSyncRouteVO"
                    }
                },
                "existing_routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.PageSyncRouteVO"
                    }
                },
                "new_routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.PageSyncRouteVO"
                    }
                },
                "retired_routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.PageSyncRouteVO"
                    }
                },
                "summary": {
                    "$ref": "#/definitions/vo.PageSyncSummaryVO"
                }
            }
        },
        "vo.PageVO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "grant_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "route_path": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}